muslim-companion/requirements/local.txt
-r base.txt

coverage==7.11.0
django-coverage-plugin==3.2.0
django-debug-toolbar==6.1.0
django-extensions==4.1
django-stubs[compatible-mypy]==5.2.7
djangorestframework-stubs==3.16.5
djlint==1.36.4
factory-boy==3.3.2
ipdb==0.13.13
mypy==1.18.2
pytest==8.4.2
pytest-django==4.11.1
pytest-sugar==1.1.1
ruff==0.14.3
sphinx==8.2.3
sphinx-autobuild==2025.8.25
watchfiles==1.1.1
werkzeug[watchdog]==3.1.3
